* C++ sucks.
git-svn-id: file:///svn/cmlib/trunk@102 25493b3f-faf5-0310-b690-0363d7aa3cd8

+++ include/gil/io/pfm.h
@@ -197,7 +197,7 @@ namespace gil {
 				double scale = ByteReverser<char>::is_little_endian() ? -1 : 1;
 
 				if (fprintf(
-						f, "P%c\n%u %u\n%f\n", magic, width, height, scale
+						f, "P%c\n%u %u\n%f\n", magic, (unsigned int)width, (unsigned int)height, scale
 					) < 0) test_and_throw(f);
 
 				Converter<ColorType, typename I::ColorType> converter;
